universal3830-common: numerous fixes
updated overlay, enabled charger, added in more selinux policies, imported props, updated recovery, enabled configstore, and fixed wrong DTBO config.

This is still a WIP. SELinux policies are still incomplete (so the builds need to be in permissive mode to function), and additional tests are needed before we can be sure that this is fully functioning.