sepolicy: fix rare RIL errors on reboot
This happened in cases where a 'distressed reboot' (ex. forced reboots or reboots due to kernel/system errors) occured. It seems to be an error with VK, as it showed SELinux denials related to vk when it happened. After applying this, such incident did not occur (even after several forced reboots via terminal, adb, and intentionally induced system error.

Hopefully, one day, we would be able to get rid of the VK s*** completely. While VK daemon itself seems to be benign (it's the VK HAL that causes all the issues), it's rather uncomfortable that we need to have this to make RIL work properly (without VK, cass keeps showing up errors. And without cass, we cannot start modem.)

Looking at the logs, it seems that A21s uses modem boot protocols that are either same or very similar to Galaxy S7, which is shown here: https://eighty-twenty.org/2020/09/10/booting-samsung-galaxy-s7-modem

Maybe this can be used as a guide to make an open-source solution to booting modem for this device, which would allow us to get rid of Sam's proprietary cbd, cassd, and vaultkeeperd.

Also, remove the auto brightness settings for the time being, as it is broken. After all, this device doesn't have the sensors for it anyway.