universal3830-common: camera: revert to Camera2 + remove notify NFC
It seems that A21s can't cope well with Snap, showing intermittent cases of unsaved/broken photos and photo sizes that are about twice of photos taken on stock.

It should be noted that additional changes are needed to Camera2 in order to change the device's capture support level to LIMITED_JPEG, or we get stuck with 12MP instead of 48MP that the device is actually capable of.

Also, the notify NFC option only caused errors, so remove it.